Move freestanding entities to documents/.
